udpate

+++ FrontEnd/src/pages/category/CategoryList.tsx
@@ -35,6 +35,7 @@ import CategoryDetail from "./CategoryDetail";
 import { Link } from "react-router-dom";
 import CategoryAdd from "./CategoryAdd";
 import { Image } from "primereact/image";
+import {checkPermission} from "@src/Service/common.ts";
 const CategoryList: React.FC = () => {
   const baseUrl = import.meta.env.VITE_BASE_URL_LOCALHOST;
   const dispatch = useDispatch();
@@ -213,7 +214,6 @@ const CategoryList: React.FC = () => {
     }).then((result) => {
       if (result.isConfirmed) {
         dispatch(setLoading(true));
-        const token = Cookies.get("token");
         $axios
           .delete(`Category/${item.Id}`)
           .then((res) => {
@@ -313,9 +313,10 @@ const CategoryList: React.FC = () => {
   const button: React.FC = () => {
     return (
       <div>
-        <div onClick={showModalAdd} className="btn btn-general">
-          Add new{" "}
-        </div>
+        {checkPermission('Category', 'create') ? (
+            <div onClick={ showModalAdd} className="btn btn-general">Add new </div>
+        ) : ''}
+
       </div>
     );
   };
@@ -476,20 +477,20 @@ const CategoryList: React.FC = () => {
                               key={crypto.randomUUID()}
                             >
                               <span
-                                className="m-2 btn-icon p-1"
-                                onClick={() => showModalDetail(item)}
+                                className={`m-2 btn-icon p-1 ${checkPermission('Category', 'read') ? '' : 'btn-disable'}`}
+                                onClick={() => checkPermission('Category', 'read') && showModalDetail(item)}
                               >
                                 <RemoveRedEyeOutlinedIcon />
                               </span>
                               <span
-                                className="m-2 btn-icon p-1"
-                                onClick={() => showModalEdit(item)}
+                                className={`m-2 btn-icon p-1 ${checkPermission('Category', 'update') ? '' : 'btn-disable'}`}
+                                onClick={() => checkPermission('Category', 'update') && showModalEdit(item)}
                               >
                                 <EditOutlinedIcon />
                               </span>
                               <span
-                                className="m-2 btn-icon btn-delete p-1"
-                                onClick={() => deleteItem(item)}
+                                className={`m-2 btn-icon p-1 ${checkPermission('Category', 'delete') ? '' : 'btn-disable'}`}
+                                onClick={() => checkPermission('Category', 'delete') && deleteItem(item)}
                               >
                                 <DeleteOutlineOutlinedIcon />
                               </span>
